requirements.txt
fastapi[all]
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
psycopg2-binary
asyncpg
python-dotenv
scikit-learn 
pandas